requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
multidict==6.0.4
notion-client==2.2.1
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2
slack-bolt==1.18.1
slack_sdk==3.26.2
sniffio==1.3.0
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
pytz